refactor: delete useContents() - 컴포넌트 내부에서 직접 호출로 변경
